Add Robot execute tests
Add a few Robot execute `Robot_exec` tests to ensure that it correctly
interprets the incoming string to function call and pass the arguments
correctly to said function call. Also, Update all Robot functions to
accept the Robot instance (which is a string of coordinates) as the
first argument, so that they update the Robot rather than generating a
new one each time.

== Notes:
- The biggest bug or concern that was troubling was not setting the
  global option for the Associative arrays; i.e. `declare -gA`; since
  when calling any function that relied on the array would fail because
  the array no longer exists outside of that context; i.e. within tests